Relocate model package
