+++ internal/app/pipeline/api/middleware/audit/middleware.go
@@ -32,7 +32,7 @@ import (
 	"github.com/banzaicloud/pipeline/internal/platform/gin/correlationid"
 	"github.com/banzaicloud/pipeline/pkg/common"
 	"github.com/banzaicloud/pipeline/secret"
-	"github.com/banzaicloud/pipeline/spotguide"
+	"github.com/banzaicloud/pipeline/src/spotguide"
 )
 
 // LogWriter instance is a Gin Middleware which logs all request data into MySQL audit_events table.
