+++ internal/ark/buckets_repository.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/banzaicloud/pipeline/auth"
 	"github.com/banzaicloud/pipeline/internal/ark/api"
-	"github.com/banzaicloud/pipeline/model"
+	"github.com/banzaicloud/pipeline/src/model"
 )
 
 // BucketsRepository descibes a repository for storing ARK backup buckets
