+++ internal/ark/deployments_model.go
@@ -20,7 +20,7 @@ import (
 	"github.com/jinzhu/gorm"
 
 	"github.com/banzaicloud/pipeline/auth"
-	"github.com/banzaicloud/pipeline/model"
+	"github.com/banzaicloud/pipeline/src/model"
 )
 
 // ClusterBackupDeploymentsModel describes an ARK deployment model
