+++ internal/cluster/model.go
@@ -22,7 +22,7 @@ import (
 	"github.com/jinzhu/gorm"
 	"github.com/sirupsen/logrus"
 
-	"github.com/banzaicloud/pipeline/model"
+	"github.com/banzaicloud/pipeline/src/model"
 )
 
 // Migrate executes the table migrations for the cluster module.
